Reorganized files, added documentation to serializers, general cleaning
