+++ src/Microsoft.Extensions.OData.Migration/Formatters/Serialization/ODataMigrationPrimitiveSerializer.cs
@@ -9,8 +9,10 @@ namespace Microsoft.Extensions.OData.Migration.Formatters.Serialization
     using Microsoft.AspNet.OData.Formatter.Serialization;
     using Microsoft.OData;
     using Microsoft.OData.Edm;
-    using System;
-    public class ODataMigrationPrimitiveSerializer : ODataPrimitiveSerializer
+    /// <summary>
+    /// Converts primitive types that serialize differently in v3 (e.g., longs are serialized with quotes in v3)
+    /// </summary>
+    internal class ODataMigrationPrimitiveSerializer : ODataPrimitiveSerializer
     {
         public ODataMigrationPrimitiveSerializer()
             : base()
